add detail router tab mechanism
* DashboardFrame -> use model-value instead of v-model.
Cause the v-model is not changeable. Generally, v-model will combined with the ref(), which is like the pointer to the variable. But in this case, computed will change the variable directly which means change the pointer, instead of the value. And this kind of action is not allowed. So I use computed that the value can directly use it to reference to current tab.
* DashBoardTable -> add handleDetail that will route to the specified id
* DashBoardTable -> fix search box bug
* dashboard store module -> make active tab as the state value
*